Final Product!
I think I did pretty well with the overall look. I was not sure how to
get "Read the reviews" to line up properly. I was also not sure how to
get the $120.00 to look the same with the small 00's at the top. But
yes, I am really happy with how it turned out.